Add most enigma logic to crypt class.wip
