+++ translated/encrypted.txt
@@ -1,1 +1,1 @@
-{:encryption=>"lojlsjuovvb", :key=>"00480", :date=>"170121"}
+{:encryption=>"ijwipeglsqo", :key=>"78077", :date=>"170121"}
